refactor(yaml_data): splitted yaml_data_schema and data_config_parsing.
